Move the analyze checks into goldeneye
goldeneye already generates and checks the dialect seeds and reserved a
place for the analysis checks, so the testcheck module folds into it:
the case loader becomes goldeneye/endtoend, reusing goldeneye's diff,
the ClickHouse analysis joins goldeneye's clickhouse package next to the
dialect generator it shares a binary with, and `check` verifies both the
committed dialect and the analyze cases. `go test ./...` in
internal/goldeneye runs the same checks.
