+++ internal/goldeneye/clickhouse/analyze.go
