+++ internal/goldeneye/cmd/goldeneye/main.go
@@ -1,13 +1,14 @@
 // Command goldeneye generates the dialect seeds under
-// internal/engine/<engine>/dialect from a live database, and checks the
-// committed ones against it.
+// internal/engine/<engine>/dialect from a live database, checks the
+// committed ones against it, and checks the analyze cases under
+// internal/endtoend/testdata against what the database itself reports.
 //
 // Usage, from internal/goldeneye:
 //
 //	go run ./cmd/goldeneye install clickhouse   # download the pinned clickhouse binary
 //	go run ./cmd/goldeneye install sqlite       # build the pinned sqlite3 shells from source
 //	go run ./cmd/goldeneye generate [engine]    # rewrite the generated files from the database
-//	go run ./cmd/goldeneye check [engine]       # compare the committed files with the database
+//	go run ./cmd/goldeneye check [engine]       # compare the committed files and analyze cases with the database
 //
 // Without an engine, generate and check cover every engine whose database
 // is available and say which ones they skipped. `go test ./...` runs the
@@ -22,10 +23,12 @@ import (
 	"io"
 	"os"
 	"runtime"
+	"strings"
 
 	"github.com/sqlc-dev/sqlc/internal/goldeneye/clickhouse"
 	"github.com/sqlc-dev/sqlc/internal/goldeneye/dialect"
 	"github.com/sqlc-dev/sqlc/internal/goldeneye/duckdb"
+	"github.com/sqlc-dev/sqlc/internal/goldeneye/endtoend"
 	"github.com/sqlc-dev/sqlc/internal/goldeneye/postgresql"
 	"github.com/sqlc-dev/sqlc/internal/goldeneye/sqlite"
 )
@@ -44,7 +47,7 @@ const usage = `usage:
   goldeneye generate [engine]
       rewrite the generated dialect files from the database, for every available engine or one
   goldeneye check [engine]
-      compare the committed dialect files with the database, for every available engine or one
+      compare the committed dialect files and analyze cases with the database, for every available engine or one
 
 engines: clickhouse, duckdb, postgresql, sqlite`
 
@@ -58,13 +61,16 @@ type engine struct {
 	version func(context.Context, string) (string, error)
 	// generate reads the dialect from the database.
 	generate func(context.Context, string) (dialect.Files, error)
+	// analyze asks the database what it reports for an analyze case, in
+	// the shape sqlc analyze prints. Nil for an engine without one yet.
+	analyze func(context.Context, string, endtoend.Case) ([]byte, error)
 }
 
 var engines = []engine{
-	{clickhouse.Engine, clickhouse.Locate, clickhouse.Version, clickhouse.Generate},
-	{duckdb.Engine, duckdb.Locate, duckdb.Version, duckdb.Generate},
-	{postgresql.Engine, postgresql.Locate, postgresql.Version, postgresql.Generate},
-	{sqlite.Engine, sqlite.Locate, sqlite.Version, sqlite.Generate},
+	{clickhouse.Engine, clickhouse.Locate, clickhouse.Version, clickhouse.Generate, clickhouse.Analyze},
+	{duckdb.Engine, duckdb.Locate, duckdb.Version, duckdb.Generate, nil},
+	{postgresql.Engine, postgresql.Locate, postgresql.Version, postgresql.Generate, nil},
+	{sqlite.Engine, sqlite.Locate, sqlite.Version, sqlite.Generate, nil},
 }
 
 // installer puts the binary an engine is read through in place, for the
@@ -204,5 +210,37 @@ func check(ctx context.Context, e engine, handle string, stderr io.Writer) error
 		return fmt.Errorf("%s does not match the database\n%s", dir, report)
 	}
 	fmt.Fprintf(stderr, "%s: ok, %d file(s) match\n", e.name, len(files))
+	return checkAnalyzeCases(ctx, e, handle, stderr)
+}
+
+// checkAnalyzeCases compares what the database reports for each of the
+// engine's analyze cases with the case's committed output.
+func checkAnalyzeCases(ctx context.Context, e engine, handle string, stderr io.Writer) error {
+	if e.analyze == nil {
+		return nil
+	}
+	cases, err := endtoend.Cases(e.name)
+	if err != nil {
+		return err
+	}
+	var report strings.Builder
+	for _, c := range cases {
+		got, err := e.analyze(ctx, handle, c)
+		if err != nil {
+			fmt.Fprintf(&report, "%s: %v\n", c.Name, err)
+			continue
+		}
+		diff, err := c.Compare(got)
+		if err != nil {
+			return err
+		}
+		if diff != "" {
+			fmt.Fprintf(&report, "%s (-committed +database)\n%s", c.Name, diff)
+		}
+	}
+	if report.Len() > 0 {
+		return fmt.Errorf("analyze cases do not match the database\n%s", report.String())
+	}
+	fmt.Fprintf(stderr, "%s: ok, %d analyze case(s) match\n", e.name, len(cases))
 	return nil
 }
