--- internal/testcheck/go.mod
@@ -1,3 +1,0 @@
-module github.com/sqlc-dev/sqlc/internal/testcheck
-
-go 1.24.0
